Daily commit for cmdline.arguments
Todays commit adds a bit of structure to the parser object (only a few skeleton features).
It is by no means complete, but it is a start for the parser project that has been added to the repository.
The template currently follows somewhat along with the structure of [argparse](https://docs.python.org/3/library/argparse.html) module from python,
but is distinct in certain ways in order to make it simpler to verify arguments by grouping these by similarity.