A long awaited update.
I've started working on the low-level backend for argument matching.
The current implementation should be very efficient, but may have some
additional optimizations to be done

Added file cmdarg_matching.cpp which contains Rcpp functions for
obtaining argument position and matching against parser arguments.

In addition I started working on the R integration side. Here it is
just a couple of functions for calling the Cpp side and for ordering
arguments.

There is only very limited amount of work left to be done in Cpp, unless
I decide to go deeper. But this is unlikely. At least I've learned a bit
about STL in c++.