comitting latest changes.
This includes finishing up the testing of parserFunction, a few comment changes in traits and the start of argument{\.h|\.cpp}.
The package workflow diagram has also been updated to include these changes, and will likely be updated with argument before argument is implemented.

Next up is implementing Argument. This will have to be able to handle many things, and a lot of care has to be taken, to ensure that it will actually do things right the first time.

Probably writing up the different things, then charting it out, figuring out which functions are needed, and then implementing each in turn.

The printer function will be needed too. I think the simplest method would be to use parserFunction for this purpose as well, but I want to allow for a string as well.

The print object is actually somewhat simple.. use std::... I know what to do. Make printer flip if it gets a function or a string. As it only ever prints when receiving "help" it does not have to be too efficient. Just flip use a nasty switch or if-else and call the appropriate function.