This is the closing of branch 0.1.0
After finishing up the bones of the package the realization of various new
problems causes a new shift in the package design.

Up till now I've tried to encapsulate the different components of a commandline argument reader
but the interaction between various components has proved to become a massive
problem for the meat of the package. Therefore the package will go through a
third design re-iteration. The main difference for this reiteration is
that I will utilize a C-style struct for holding shared data.
This C-style struct will only be stored in the argument class private members,
while other components will only store a reference to the object.

This is somewhat destroying the original idea of using OOP for handling the various components,
but the problem with this approach is a large amount of duplicate initializers.

For once, I think this is the best approach to solve some of the problems I've been grumbling about.